backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-pymongo==2.3.0
flask-mail==0.9.1
pymongo==4.7.3
python-dotenv==1.0.1
bcrypt==4.1.3
requests==2.32.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dateutil==2.8.2
numpy==1.24.3
opencv-python==4.8.1.78
jinja2==3.1.2
schedule==1.2.2
Pillow==12.2.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
urllib3==1.26.18
